REFACTOR suggestArguments into own file
